+++ src/main/java/ch/danielt/akw/effect/RadiationEffect.java
@@ -1,0 +1,10 @@
+package ch.danielt.akw.effect;
+
+import net.minecraft.entity.effect.StatusEffect;
+import net.minecraft.entity.effect.StatusEffectCategory;
+
+public class RadiationEffect extends StatusEffect {
+    public RadiationEffect() {
+        super(StatusEffectCategory.HARMFUL, 0x39D353);
+    }
+}